Small project refactor
