+++ src/index.ts
@@ -1,4 +1,4 @@
-import Number from './models/schemas/number'
-import String from './models/schemas/string'
+import Number from './models/schemas/numberSchema'
+import String from './models/schemas/stringSchema'
 
 export default { String, Number }
