+++ src/models/requirements/string/stringRequirement.ts
@@ -2,12 +2,6 @@ import type { RequirementValidationError } from '../../../types/ValidationResult
 import Requirement from '../reqirement'
 
 export default class StringRequirement extends Requirement {
-  /**
-   * Validate that the value is of type string.
-   *
-   * @param value Value to validate.
-   * @returns True if the value is a string, false otherwise.
-   */
   public validate(value: any): RequirementValidationError[] {
     if (typeof value === 'string') return []
